20-06-2018

+++ CapaLogicaNegocio/basic/registrarvManager.vb
@@ -13,7 +13,7 @@ Namespace BLL
         End Function
 
         Public Shared Function GetList(ByVal vidalmacen As Integer, ByVal vmes As Integer, ByVal vanio As Integer, ByVal vproceso As String,
-                                       ByVal varrrol As String, ByVal vfilas As Long, ByVal vcodigo_per As Integer) As DataTable
+                                       ByVal varrrol As String, ByVal vfilas As Integer, ByVal vcodigo_per As Integer) As DataTable
             Return registrarVBD.GetList(vidalmacen, vmes, vanio, vproceso, varrrol, vfilas, vcodigo_per)
         End Function
